GUEST USER & LOGGEDIN USER

+++ lib/model/restaurant/restaurantlist.dart
@@ -10,7 +10,7 @@ class Restaurantlist {
   dynamic officialWebsite;
   dynamic contactNumber;
   dynamic email;
-  int? distance;
+  double? distance;
   dynamic rating;
   int? ratingCount;
   String? latitude;
@@ -51,7 +51,7 @@ class Restaurantlist {
       officialWebsite: data['officialWebsite'] as dynamic,
       contactNumber: data['contactNumber'] as dynamic,
       email: data['email'] as dynamic,
-      distance: data['distance'] as int?,
+      distance: data['distance'] as double?,
       rating: data['rating'] as dynamic,
       ratingCount: data['ratingCount'] as int?,
       latitude: data['latitude'] as String?,
@@ -100,7 +100,7 @@ class Restaurantlist {
     dynamic officialWebsite,
     dynamic contactNumber,
     dynamic email,
-    int? distance,
+    double? distance,
     dynamic rating,
     int? ratingCount,
     String? latitude,
